fix: 🐛 critical circular dependencies
