+++ src/components/Collapse/style/Collapse.style.ts
@@ -1,3 +1,3 @@
-import { styled } from '../../..';
+import { styled } from '../../../styled';
 
 export const StyledCollapse = styled.div``;
